requirements.txt
fastapi
uvicorn
langchain
faiss-cpu
pydantic
requests
python-multipart
langchain-community
langchain-huggingface
sentence-transformers